celltyping/pyproject.toml
[project]
name = "immuvis-celltyping"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anndata>=0.12.10",
    "bio>=1.8.1",
    "comet-ml>=3.57.4",
    "delayedarray>=0.6.2",
    "einops>=0.8.2",
    "fair-esm>=2.0.0",
    "huggingface-hub>=1.5.0",
    "loguru>=0.7.3",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "omegaconf>=2.3.0",
    "opencv-python>=4.13.0.92",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "rds2py==0.4.0",
    "requests>=2.32.5",
    "ruamel-yaml>=0.19.1",
    "safetensors>=0.7.0",
    "scikit-image>=0.26.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "tifffile>=2026.2.16",
    "timm>=1.0.25",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "tqdm>=4.67.3",
    "umap-learn>=0.5.11",
    "xformers>=0.0.29.post3",
]
# uv pip install https://github.com/mjun0812/flash-attention-prebuild-wheels/releases/download/v0.7.16/flash_attn-2.8.3%2Bcu124torch2.6-cp312-cp312-linux_x86_64.whl


[tool.uv.sources]
torch = [{ index = "pytorch-cu124" }]
torchvision = [{ index = "pytorch-cu124" }]
torchaudio = [{ index = "pytorch-cu124" }]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
